create symlinks only for .files

+++ bootstrap.sh
@@ -4,19 +4,21 @@ function doIt() {
 	local files=`git ls-tree --name-only HEAD`
 
 	for i in $files; do
-		rm -rf "$HOME/$i"
-    	ln -s "$PWD/$i" "$HOME/$i"
+		if [[ $i == \.* ]]; then
+			rm -rf "$HOME/$i"
+			ln -s "$PWD/$i" "$HOME/$i"
+		fi
 	done
 
 	local zsh=`grep -m 1 zsh /etc/shells`
 	if [[ "$SHELL" = *zsh ]]; then
-		`$zsh && source "$HOME/.zshrc"`
+		#`exec zsh`
 	else
 		if [[ -n $zsh ]]; then
 			read -n 1 -p "Change your default shell to zsh? (y/n) "
 			echo
 			chsh -s /bin/sh
-			`$zsh`
+			#`exec $zsh`
 		else
 			echo "Please install zsh and make it default shell"
 		fi
